chore: promote encodeAtlasStateID for further use
